Enable /reset endpoint only on those environments which are not production

+++ app/app.rb
@@ -124,16 +124,19 @@ error OrderNotFoundError do |e|
   { error: e.message }.to_json
 end
 
-post '/reset' do
-  OrderRepository.new.delete_all
-  ClientRepository.new.delete_all
-  DeliveryRepository.new.delete_all
-  status 200
-end
-
 def extract_first_error(entity)
   return '' if entity.errors.empty?
 
   # Ex: entity.errors.messages = [:symbol, ["the error"]]
   entity.errors.messages.first[1].first
+end
+
+# test environment methods
+if settings.environment != :production
+  post '/reset' do
+    OrderRepository.new.delete_all
+    ClientRepository.new.delete_all
+    DeliveryRepository.new.delete_all
+    status 200
+  end
 end
